Make footer responsive for mobile
