+++ features/ui/index.ts
@@ -2,3 +2,4 @@ export * from "./badge";
 export * from "./button";
 export * from "./page-container";
 export * from "./sidebar-navigation";
+export * from "./footer";
