+++ features/ui/page-container/page-container.tsx
@@ -1,9 +1,8 @@
 import React from "react";
 import Head from "next/head";
 import styled from "styled-components";
-import { SidebarNavigation } from "@features/ui";
+import { SidebarNavigation, Footer } from "@features/ui";
 import { color, displayFont, space, breakpoint, textFont } from "@styles/theme";
-import Link from "next/link";
 
 type PageContainerProps = {
   children: React.ReactNode;
@@ -57,51 +56,6 @@ const Info = styled.div`
   ${textFont("md", "regular")}
 `;
 
-// Footer
-const Footer = styled.footer`
-  background-color: ${color("gray", 50)};
-  height: 3.7rem;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding-inline: ${space(8)};
-`;
-
-const Version = styled.p`
-  color: ${color("gray", 400)};
-`;
-
-const Nav = styled.nav`
-  margin-right: ${space(16)};
-`;
-
-const List = styled.ul`
-  margin: 0;
-  padding: 0;
-  position: inline-block;
-  display: flex;
-  gap: ${space(6)};
-  list-style-type: none;
-  width: 18rem;
-  /* border: 1px solid red; */
-  /* justify-content: space-between; */
-`;
-
-const ListItem = styled(Link)`
-  position: block;
-`;
-
-const Anchor = styled.a`
-  display: block;
-  color: ${color("gray", 500)};
-  font: ${textFont("sm", "regular")};
-`;
-
-const MenuIcon = styled.img`
-  display: block;
-  width: ${space(6)} - 0.25;
-`;
-
 export function PageContainer({ children, title, info }: PageContainerProps) {
   return (
     <Container>
@@ -118,26 +72,7 @@ export function PageContainer({ children, title, info }: PageContainerProps) {
           <Info>{info}</Info>
           {children}
         </ContentContainer>
-        <Footer>
-          <Version>Version: 14.5.1</Version>
-          <Nav>
-            <List>
-              <ListItem href="/dashboard">
-                <Anchor>Docs</Anchor>
-              </ListItem>
-              <ListItem href="/dashboard">
-                <Anchor>API</Anchor>
-              </ListItem>
-              <ListItem href="/dashboard">
-                <Anchor>Help</Anchor>
-              </ListItem>
-              <ListItem href="/dashboard">
-                <Anchor>Community</Anchor>
-              </ListItem>
-            </List>
-          </Nav>
-          <MenuIcon src="/icons/logo-small.svg" alt="logo" />
-        </Footer>
+        <Footer />
       </Main>
     </Container>
   );
